requirements.txt
# cryptography is required for heroku deployment
coverage==3.6
django==2.0.13
dj-database-url==0.2.1
gunicorn==0.17.4
psycopg2==2.8.3
dj-static==0.0.6
django-reversion==3.0.1
#django-axes==1.4.0
ffs==0.0.8.2
letter==0.5
requests==2.22.0
djangorestframework==3.10.2
django-compressor==2.2
python-dateutil==2.8.0
django-celery==3.2.2
celery==3.1.25
whitenoise==5.1.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1